+++ app.territoiresentransitions.react/src/app/pages/collectivite/ParcoursLabellisation/useCycleLabellisation.ts
@@ -1,11 +1,10 @@
-import {useQuery} from 'react-query';
 import {TLabellisationParcours} from 'app/pages/collectivite/ParcoursLabellisation/types';
-import {supabaseClient} from 'core-logic/api/supabase';
 import {useCollectiviteId} from 'core-logic/hooks/params';
 import {useIsAuditeur} from '../Audit/useAudit';
 import {useCarteIdentite} from '../PersoReferentielThematique/useCarteIdentite';
 import {usePreuves} from 'ui/shared/preuves/Bibliotheque/usePreuves';
 import {TPreuveLabellisation} from 'ui/shared/preuves/Bibliotheque/types';
+import {useLabellisationParcours} from './useLabellisationParcours';
 
 // données du cycle de labellisation/audit actuel d'une collectivité
 export type TCycleLabellisation = {
@@ -33,13 +32,7 @@ export const useCycleLabellisation = (
   const identite = useCarteIdentite(collectivite_id);
 
   // charge les données du parcours
-  const {data: parcoursList} = useQuery(
-    ['labellisation_parcours', collectivite_id],
-    () => fetchParcours(collectivite_id)
-  );
-
-  // extrait le parcours correspondant au référentiel courant
-  const parcours = getReferentielParcours(parcoursList, referentiel);
+  const parcours = useLabellisationParcours({collectivite_id, referentiel});
   const {completude_ok, rempli} = parcours || {};
 
   // charge les documents de labellisation
@@ -89,47 +82,4 @@ const getParcoursStatus = (
     return 'demande_envoyee';
   }
   return 'non_demandee';
-};
-
-// charge les parcours (eci/cae) de labellisation d'une collectivité donnée
-const fetchParcours = async (
-  collectivite_id: number | null
-): Promise<TLabellisationParcours[] | null> => {
-  if (!collectivite_id) {
-    return null;
-  }
-
-  const {data, error} = await supabaseClient
-    .rpc('labellisation_parcours', {
-      collectivite_id,
-    })
-    .select();
-
-  if (error || !data) {
-    return null;
-  }
-  return data.map(d => ({
-    ...d,
-    collectivite_id, // on ajoute l'id qui n'est pas redonné par la vue
-  })) as TLabellisationParcours[];
-};
-
-export const getReferentielParcours = (
-  parcoursList: TLabellisationParcours[] | null | undefined,
-  referentiel: string | null
-) => {
-  const parcours: TLabellisationParcours | undefined = parcoursList?.find(
-    p => p.referentiel === referentiel
-  );
-
-  if (!parcours) {
-    return null;
-  }
-
-  const {criteres_action} = parcours;
-  return {
-    ...parcours,
-    // trie les critères action
-    criteres_action: criteres_action.sort((a, b) => a.prio - b.prio),
-  };
 };
